% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{celsius_from_kelvin}
\alias{celsius_from_kelvin}
\title{Convert Kelvin to degrees Celsius}
\usage{
celsius_from_kelvin(kelvin)
}
\arguments{
\item{kelvin}{Numeric vector of temperatures in Kelvin.}
}
\value{
A numeric vector of temperatures in degrees Celsius.
}
\description{
Convert Kelvin to degrees Celsius
}
\examples{
celsius_from_kelvin(c(273.15, 295.15))
}
