% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwi.R
\name{mwi_fhft}
\alias{mwi_fhft}
\title{Temperature-and-humidity variant of the index (FHFT)}
\usage{
mwi_fhft(temperature, humidity)
}
\arguments{
\item{temperature}{Numeric vector of air temperatures in degrees Celsius.}

\item{humidity}{Numeric vector of relative humidities in percent (0-100).}
}
\value{
A numeric vector of index values in \verb{[0, 1]}.
}
\description{
The Mosquito Weather Index with the wind term omitted, that is the product of
\code{\link[=mwi_humidity]{mwi_humidity()}} and \code{\link[=mwi_temperature]{mwi_temperature()}} only. Useful when no wind
observations are available, for example when working with temperature and
humidity loggers.
}
\examples{
mwi_fhft(temperature = 22, humidity = 70)

# Equal to mwi() whenever wind is below the cut-off
identical(
  mwi_fhft(22, 70),
  mwi(22, 70, wind_speed = 0)
)

}
