% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwi-package.R
\docType{package}
\name{mwi-package}
\alias{mwi-package}
\title{mwi: Compute the Mosquito Weather Index}
\description{
Computes the Mosquito Weather Index (MWI), an operational index that combines air temperature, relative humidity and wind speed into a single number between 0 and 1 summarising how favourable the weather is for adult mosquito activity, and maps that number onto five ordered activity classes. The index was developed by the Benaki Phytopathological Institute and the National Observatory of Athens under the LIFE CONOPS project and has been published operationally since 2014. Because the index is a non-linear function of its three inputs, it must be evaluated at sub-daily resolution before being aggregated over time; helpers are provided that enforce that order of operations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/E4Warning/mwi}
  \item \url{https://e4warning.github.io/mwi/}
  \item Report bugs at \url{https://github.com/E4Warning/mwi/issues}
}

}
\author{
\strong{Maintainer}: John R. B. Palmer \email{john.palmer@upf.edu} (\href{https://orcid.org/0000-0002-2648-7860}{ORCID})

Authors:
\itemize{
  \item Antonios Michaelakis (\href{https://orcid.org/0000-0002-3075-5020}{ORCID})
  \item Georgios Balatsos (\href{https://orcid.org/0000-0002-3973-068X}{ORCID})
  \item Vasileios Karras
  \item Dimitrios Papachristos
}

}
\keyword{internal}
